refactor(activitypub): move local auto-repost logic into member service
The initial fix for multi-hop auto-repost put performLocalAutoReposts()
and its BFS helper directly on ActivityPubInterface, which violates the
thin-delegation pattern established for domain interface classes.
Interface methods should delegate to services, not contain DB access,
cycle-prevention logic, or event bus emission.

Moves both methods verbatim into ActivityPubMemberService (which
already owns shareEvent, followCalendar, and the other member-side
federation actions, and already has eventBus + calendarService in
scope). The interface now exposes performLocalAutoReposts() as a
one-line delegation.

Drops the interface's private eventBus field and logger import, which
were added solely for the now-moved methods.

Pre-existing drift in the same file (direct entity access in several
other methods) is tracked separately in pv-3vx9.

No behavior change; all auto-repost integration tests pass.
